Add IsGardenContainer function
Add function `IsGardenContainer` to detect whether the current process
is started in a Garden Container process tree, which is helpful for terminal
related details like the terminal width. Inside a Garden Container, you
cannot rely on the terminal width and height being reported as it can be set
to any values during Garden set-up.