Refactor time info text display and introduce timeout
Introduce a timeout, which will be used to (A) stop the progress indicator
after it is reached and (B) to smoothly blend the designed text color into
red to less time is on the clock.

Refactor the time info text function, which is used to created the info text
on the far right side of the progress indicator. A user can supply a custom
function to create whatever text is needed and whatever color is wanted.